Implement LRU cache for blacklist
Although the local dict cache works fine, there is a lot of cache invalidation and others that would need to be worried about. The cache would grow without limit, so an LRU cache is used to bind that limit in order to evict stale data. This also helps with performance as it's constantly hitting the cache instead of checking a local dict, which the contents can be wiped away with a GC call